L15-structuralPatterns/homework/src/test/java/ru/otus/processor/homework/ProcessorReturnExcEventSecondTest.java
package ru.otus.processor.homework;

import org.junit.jupiter.api.Test;
import ru.otus.model.Message;
import ru.otus.model.ObjectForMessage;

import java.time.LocalDateTime;
import java.time.Month;
import java.util.ArrayList;

import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
import static org.junit.jupiter.api.Assertions.assertThrows;
import static org.junit.jupiter.api.Assertions.assertTrue;

public class ProcessorReturnExcEventSecondTest {

    @Test
    void processorTest(){

        var processor = new ProcessorReturnExcEventSecond(() -> LocalDateTime.of(2023, Month.JUNE, 04, 15, 32, 40));

        var id = 100L;
        var data = "33";
        var field13 = new ObjectForMessage();
        var field13Data = new ArrayList<String>();
        field13Data.add(data);
        field13.setData(field13Data);

        var message = new Message.Builder(id)
                .field10("field10")
                .field13(field13)
                .build();

        Exception exception = assertThrows(RuntimeException.class, () -> {
            processor.process(message);
        });
        String expectedMessage = "Event Second";
        String actualMessage = exception.getMessage();
        assertTrue(actualMessage.contains(expectedMessage));

    }
}
